- merged basicUser and added a boolean to signify premium settings - deleted premiumUser and respective unit test cases file
