+++ src/main/java/entities/basicUser.java
@@ -1,10 +1,17 @@
 package entities;
 
-public class basicUser extends loginInformation{
+import usecases.EditUser;
+
+import java.util.Date;
+
+public class basicUser extends loginInformation {
     private String ID;
     private String email;
     private String phoneNumber;
     private String classType;
+    private boolean upgraded;
+    private Date renewalDate;
+    private int appRating;
 
     public basicUser(){
 
@@ -15,6 +22,17 @@ public class basicUser extends loginInformation{
         this.email = email;
         this.phoneNumber = phoneNumber;
         this.classType = c;
+        this.upgraded = false;
+    }
+
+    public basicUser(String id, String email, String phoneNumber, String c, Boolean upgradeAccount){
+        this.ID = id;
+        this.email = email;
+        this.phoneNumber = phoneNumber;
+        this.classType = c;
+        this.upgraded = upgradeAccount;
+        this.renewalDate = renewalDate;
+        this.appRating = appRating;
     }
 
     public String getID() {
@@ -47,6 +65,22 @@ public class basicUser extends loginInformation{
 
     public void setClassType(String classType) {
         this.classType = classType;
+    }
+
+    public Date getRenewalDate() {
+        return renewalDate;
+    }
+
+    public void setRenewalDate(Date renewalDate) {
+        this.renewalDate = renewalDate;
+    }
+
+    public int getAppRating() {
+        return appRating;
+    }
+
+    public void setAppRating(int appRating) {
+        this.appRating = appRating;
     }
