--- src/test/java/premiumUserTest.java
@@ -1,2 +1,0 @@
-public class premiumUserTest {
-}
